[Fix #1828]check whether has permission to download udf file or delete udf function (#1858)
* fix issue 1828:get the udf resource path error when create udf function

* update grantResources

* first verify whether udf resource is bound by udf function

* update grantResources

* update testListAuthorizedUdfFunc

* update getUserInfo in order to run success

* check whether has permission to download udf file or delete udf file

* update listAuthorizedResourceById in ResourceMapper.xml

+++ dolphinscheduler-api/src/main/java/org/apache/dolphinscheduler/api/service/UdfFuncService.java
@@ -301,13 +301,24 @@ public class UdfFuncService extends BaseService{
     /**
      * delete udf function
      *
-     * @param id udf function id
+     * @param id        udf function id
+     * @param loginUser login user
      * @return delete result code
      */
     @Transactional(rollbackFor = Exception.class)
-    public Result delete(int id) {
+    public Result delete(User loginUser,int id) {
         Result result = new Result();
-        
+
+        UdfFunc udfFunc = udfFuncMapper.selectUdfById(id);
+        if(udfFunc == null){
+            putMsg(result,Status.UDF_FUNCTION_NOT_EXIST);
+            return result;
+        }
+
+        if (!hasPerm(loginUser,udfFunc.getUserId())){
+            putMsg(result,Status.USER_NO_OPERATION_PERM);
+            return result;
+        }
         udfFuncMapper.deleteById(id);
         udfUserMapper.deleteByUdfFuncId(id);
         putMsg(result, Status.SUCCESS);
